Only override schema on sum, min and max by definition will not overflow

+++ blaze/compute/tests/test_numpy_compute.py
@@ -189,7 +189,7 @@ def test_summary_on_ndarray():
 
     result = compute(summary(total=a.sum(), min=a.min(), keepdims=True), ax)
     expected = np.array([(ax.min(), ax.sum())],
-                        dtype=[('min', 'float64'), ('total', 'float64')])
+                        dtype=[('min', 'float32'), ('total', 'float64')])
     assert result.ndim == ax.ndim
     assert eq(expected, result)
